Feat/saved cards

+++ src/Components/Card/Card.js
@@ -1,3 +1,4 @@
+import PropTypes from "prop-types";
 import "./Card.css";
 
 const handleSave = (props) => {
@@ -5,30 +6,16 @@ const handleSave = (props) => {
   const saved = localStorage.getItem(localKey);
   console.log("saved", saved);
 
-  [
-    {
-      platform: "amazonStore",
-      title: "dashstar* (VIP)",
-      artistName: "Knock2",
-      thumbnailUrl: "https://m.media-amazon.com/images/I/51nxztXYmoL.jpg",
-      id: "B0BRMQPH4V",
-      link: "https://amazon.com/dp/B0BRMQPH4V",
-    },
-    {
-      platform: "amazonMusic",
-      title: "dashstar* (VIP)",
-      artistName: "Knock2",
-      thumbnailUrl: "https://m.media-amazon.com/images/I/51nxztXYmoL.jpg",
-      id: "B0BRMQPH4V",
-      link: "https://music.amazon.com/albums/B0BRMDR92C?trackAsin=B0BRMQPH4V",
-    },
-  ];
 
   if (saved) {
     const savedArr = JSON.parse(saved);
-    const updatedSavedArr = [...savedArr, props];
-    // TODO: add a check for current item before saving
-    localStorage.setItem(localKey, JSON.stringify(updatedSavedArr));
+    const isItemAlreadySaved = savedArr.some(savedObj => savedObj.id === props.id && savedObj.platform === props.platform);
+    if (!isItemAlreadySaved) {
+      const updatedSavedArr = [...savedArr, props];
+      localStorage.setItem(localKey, JSON.stringify(updatedSavedArr));
+    } else {
+      console.warn('Item already exists in local storage:', props);
+    }
   } else {
     localStorage.setItem(localKey, JSON.stringify([props]));
   }
@@ -54,3 +41,12 @@ const Card = (props) => {
 };
 
 export default Card;
+
+Card.propTypes = {
+  platform: PropTypes.string.isRequired,
+  title: PropTypes.string.isRequired,
+  artistName: PropTypes.string.isRequired,
+  id: PropTypes.string.isRequired,
+  thumbnailUrl: PropTypes.string.isRequired,
+  link: PropTypes.string.isRequired,
+};
